[Feat/DetailPage] 게시글 찜 코드 분리
