+++ src/components/product-detail-page/pdp-right/elements/SmallUI/ToolBoxForWriter.tsx
@@ -3,9 +3,9 @@
 // import { useState } from "react";
 import Link from "next/link";
 import styled from "styled-components";
-import { useState } from "react";
+// import { useState } from "react";
 import { AiOutlineShareAlt, AiOutlineEdit } from "react-icons/ai";
-import { DeleteModal } from "../deletePost";
+import { DeleteModal } from "../DeletePost";
 
 const ToolBoxForWriterBox = styled.div`
   display: flex;
